* As a AU, I can see that view is updating relevant data each 1 sec.

+++ app/src/main/java/com/prituladima/android/revolut/model/Repository.java
@@ -4,7 +4,6 @@ import com.prituladima.android.revolut.RevolutApplication;
 import com.prituladima.android.revolut.model.api.CurrencyAPI;
 import com.prituladima.android.revolut.model.db.HawkLocalStorage;
 import com.prituladima.android.revolut.model.dto.Currency;
-import com.prituladima.android.revolut.model.dto.RemoteCurrencyDTO;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,11 +12,8 @@ import java.util.Map;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import retrofit2.http.Path;
 import rx.Observable;
-import rx.Scheduler;
 import rx.android.schedulers.AndroidSchedulers;
-import rx.functions.Func1;
 import rx.schedulers.Schedulers;
 
 @Singleton
@@ -33,7 +29,7 @@ public class Repository {
         RevolutApplication.getInjector().inject(this);
     }
 
-    public Observable<List<Currency>> updateCurrencies(String base, int amount) {
+    public Observable<List<Currency>> updateCurrencies(String base, Double amount) {
         return currencyAPI.getCurrencies(base, amount)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
